fastapi-todo/app/schemas.py
from pydantic import BaseModel, EmailStr
from typing import List, Optional


class TodoBase(BaseModel):
    title: str
    description: Optional[str] = None
    completed: bool = False


class TodoCreate(TodoBase):
    pass


class TodoUpdate(TodoBase):
    pass


class Todo(TodoBase):
    id: int
    owner_id: int

    class Config:
        orm_mode = True


class UserBase(BaseModel):
    username: str
    email: EmailStr


class UserCreate(UserBase):
    password: str


class User(UserBase):
    id: int
    is_active: bool
    is_admin: bool
    todos: List[Todo] = []

    class Config:
        orm_mode = True

class Token(BaseModel):
    access_token: str
    token_type: str